requirements.txt
python-chess==0.19.0
peewee==2.10.0
tqdm==4.29.1
